% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az_vmss_resource.R
\docType{class}
\name{az_vmss_resource}
\alias{az_vmss_resource}
\title{Virtual machine scaleset (cluster) resource class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
az_vmss_resource
}
\description{
Class representing a virtual machine scaleset resource.
}
\keyword{datasets}
