% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az_vmss_template.R
\docType{class}
\name{az_vmss_template}
\alias{az_vmss_template}
\title{Virtual machine scaleset (cluster) template class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
az_vmss_template
}
\description{
Class representing a virtual machine scaleset template.
}
\keyword{datasets}
