% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_json.R
\name{build_template_definition.vm_config}
\alias{build_template_definition.vm_config}
\alias{build_template_definition.vmss_config}
\alias{build_template_parameters.vm_config}
\alias{build_template_parameters.vmss_config}
\title{Build template definition and parameters}
\usage{
\method{build_template_definition}{vm_config}(config, ...)

\method{build_template_definition}{vmss_config}(config, ...)

\method{build_template_parameters}{vm_config}(config, name, login_user,
  size, ...)

\method{build_template_parameters}{vmss_config}(config, name, login_user,
  size, instances, ...)
}
\arguments{
\item{config}{An object of class \code{vm_config} representing a virtual machine deployment.}

\item{...}{For \code{build_template_parameters}, any additional parameters that the VM deployment template may need. Unused for \code{build_template_definition}.}

\item{name}{The VM name. Will also be used for the domain name label.}

\item{login_user}{An object of class \code{user_config} representing the login details for the admin user account on the VM.}

\item{size}{The VM size.}

\item{instances}{For \code{vmss_config}, the number of (initial) instances in the VM scaleset.}
}
\description{
Build template definition and parameters
}
