% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_methods.R
\name{defunct}
\alias{defunct}
\alias{get_vm_cluster}
\alias{create_vm_cluster}
\alias{delete_vm_cluster}
\title{Defunct methods}
\description{
Defunct methods
}
\section{Usage}{
\preformatted{get_vm_cluster(...)
create_vm_cluster(...)
delete_vm_cluster(...)
}

These methods for the \code{az_subscription} and \code{az_resource_group} classes are defunct in AzureVM 2.0. To work with virtual machine clusters, call the \link{get_vm_scaleset}, \link{create_vm_scaleset} and \link{delete_vm_scaleset} methods instead.
}

