% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_resource_config.R
\name{ip_config}
\alias{ip_config}
\title{Public IP address configuration}
\usage{
ip_config(type = "basic", dynamic = if (type == "basic") TRUE else
  FALSE, ipv6 = FALSE, domain_name = "[parameters('vmName')]", ...)
}
\arguments{
\item{type}{The SKU of the IP address resource: "basic" or "standard".}

\item{dynamic}{Whether the IP address should be dynamically or statically allocated. Note that the standard SKU only supports standard allocation.}

\item{ipv6}{Whether to create an IPv6 address. The default is IPv4.}

\item{domain_name}{The domain name label to associate with the address.}
}
\description{
Public IP address configuration
}
