% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az_vm_template.R
\name{is_vm}
\alias{is_vm}
\alias{is_vm_template}
\alias{is_vm_resource}
\title{Is an object an Azure VM}
\usage{
is_vm(object)

is_vm_template(object)

is_vm_resource(object)
}
\arguments{
\item{object}{an R object.}
}
\value{
These functions return TRUE for an object representing a VM template deployment (for \code{is_vm} and \code{is_vm_template}) or resource (for \code{is_vm_resource}), FALSE otherwise.
}
\description{
Is an object an Azure VM
}
