% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az_vm_template.R
\name{is_vm_template}
\alias{is_vm_template}
\title{Is an object an Azure VM template}
\usage{
is_vm_template(object)
}
\arguments{
\item{object}{an R object.}
}
\value{
A boolean.
}
\description{
Is an object an Azure VM template
}
\details{
This function returns TRUE only for an object representing a VM template deployment. In particular, it returns FALSE for a raw VM resource.
}
