% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_config.R
\docType{data}
\name{lb_rule_ssh}
\alias{lb_rule_ssh}
\alias{lb_rule_http}
\alias{lb_rule_https}
\alias{lb_rule_rdp}
\alias{lb_rule_jupyter}
\alias{lb_rule_rstudio}
\alias{lb_rule_mssql}
\alias{lb_rule_mssql_browser}
\alias{lb_probe_ssh}
\alias{lb_probe_http}
\alias{lb_probe_https}
\alias{lb_probe_rdp}
\alias{lb_probe_jupyter}
\alias{lb_probe_rstudio}
\alias{lb_probe_mssql}
\alias{lb_probe_mssql_browser}
\title{Load balancing rules}
\format{Objects of class \code{lb_rule} and \code{lb_probe}.}
\usage{
lb_rule_ssh

lb_rule_http

lb_rule_https

lb_rule_rdp

lb_rule_jupyter

lb_rule_rstudio

lb_rule_mssql

lb_rule_mssql_browser

lb_probe_ssh

lb_probe_http

lb_probe_https

lb_probe_rdp

lb_probe_jupyter

lb_probe_rstudio

lb_probe_mssql

lb_probe_mssql_browser
}
\description{
Load balancing rules
}
\details{
Some predefined load balancing objects, for commonly-used ports.
\itemize{
\item HTTP: TCP port 80
\item HTTPS: TCP port 443
\item JupyterHub: TCP port 8000
\item RDP: TCP port 3389
\item RStudio Server: TCP port 8787
\item SSH: TCP port 22
\item SQL Server: TCP port 1433
\item SQL Server browser service: TCP port 1434
}
}
\keyword{datasets}
