% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_methods.R
\name{list_vm_sizes}
\alias{list_vm_sizes}
\title{List available VM sizes}
\usage{
list_vm_sizes(location, name_only=FALSE)
}
\arguments{
\item{location}{The location/region for which to obtain available VM sizes.}

\item{name_only}{Whether to return only a vector of names, or all information on each VM size.}
}
\value{
If \code{name_only} is TRUE, a character vector of names, suitable for passing to \code{create_vm}. If FALSE, a data frame containing the following information for each VM size: the name, number of cores, OS disk size, resource disk size, memory, and maximum data disks.
}
\description{
Method for the \link[AzureRMR:az_subscription]{AzureRMR::az_subscription} class.
}
\seealso{
\link{create_vm}
}
