% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_config.R
\docType{data}
\name{nsg_rule_allow_ssh}
\alias{nsg_rule_allow_ssh}
\alias{nsg_rule_allow_http}
\alias{nsg_rule_allow_https}
\alias{nsg_rule_allow_rdp}
\alias{nsg_rule_allow_jupyter}
\alias{nsg_rule_allow_rstudio}
\title{Network security rules}
\format{An object of class \code{nsg_rule_config} of length 2.}
\usage{
nsg_rule_allow_ssh

nsg_rule_allow_http

nsg_rule_allow_https

nsg_rule_allow_rdp

nsg_rule_allow_jupyter

nsg_rule_allow_rstudio
}
\description{
Network security rules
}
\section{Usage}{
\preformatted{nsg_rule_allow_http
nsg_rule_allow_https
nsg_rule_allow_jupyter
nsg_rule_allow_rdp
nsg_rule_allow_rstudio
nsg_rule_allow_ssh
}
}

\section{Details}{

These are the default network security rules provided by AzureVM. They are setup to allow specific ports through the Azure firewall. You should only allow the ports that you need.
\itemize{
\item HTTP: TCP port 80
\item HTTPS: TCP port 443
\item JupyterHub: TCP port 8000
\item RDP: TCP port 3389
\item RStudio Server: TCP port 8787
\item SSH: TCP port 22
}
}

\keyword{datasets}
