% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVM.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{build_template_definition}
\alias{build_template_parameters}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{AzureRMR}{\code{\link[AzureRMR:build_template]{build_template_definition}}, \code{\link[AzureRMR:build_template]{build_template_parameters}}}
}}

