% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_methods.R
\name{delete_vm}
\alias{delete_vm}
\title{Delete virtual machine}
\usage{
delete_vm(name, confirm=TRUE, free_resources=TRUE)
delete_vm(name, confirm=TRUE, free_resources=TRUE, resource_group=name)
}
\arguments{
\item{name}{The VM name.}

\item{confirm}{Whether to confirm the delete.}

\item{free_resources}{If the VM is a deployed template, whether to free all resources created during the deployment process.}

\item{resource_group}{For the \code{AzureRMR::az_subscription} method, the resource group containing the VM.}
}
\description{
Method for the \link[AzureRMR:az_subscription]{AzureRMR::az_subscription} and \link[AzureRMR:az_resource_group]{AzureRMR::az_resource_group} classes.
}
\details{
If the VM is of class \link{az_vm_template} and was created in exclusive mode, this method deletes the entire resource group containing the VM. This automatically frees all resources that were created during the deployment process. Otherwise, if \code{free_resources=TRUE}, it manually deletes each individual resource in turn. This is done synchronously (the method does not return until the deletion is complete) to allow for dependencies.

If the VM is of class \link{az_vm_resource}, this method only deletes the VM resource itself, not any other resources it may depend on.
}
\seealso{
\link{create_vm}, \link{az_vm_template}, \link{az_vm_resource},
\link[AzureRMR:az_subscription]{AzureRMR::az_subscription}, \link[AzureRMR:az_resource_group]{AzureRMR::az_resource_group}
}
