% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_vm.R
\name{is_vm}
\alias{is_vm}
\alias{is_vm_template}
\alias{is_vm_resource}
\alias{is_vm_scaleset}
\alias{is_vm_scaleset_template}
\alias{is_vm_scaleset_resource}
\title{Is an object an Azure VM}
\usage{
is_vm(object)

is_vm_template(object)

is_vm_resource(object)

is_vm_scaleset(object)

is_vm_scaleset_template(object)

is_vm_scaleset_resource(object)
}
\arguments{
\item{object}{an R object.}
}
\value{
\code{is_vm} and \code{is_vm_template} return TRUE for an object representing a virtual machine deployment (which will include other resources besides the VM itself).

\code{is_vm_resource} returns TRUE for an object representing the specific VM resource.

\code{is_vm_scaleset} and \code{is_vm_scaleset_template} return TRUE for an object representing a VM scaleset deployment.

\code{is_vm_scaleset_resource} returns TRUE for an object representing the specific VM scaleset resource.
}
\description{
Is an object an Azure VM
}
\seealso{
\link{create_vm}, \link{create_vm_scaleset}, \link{az_vm_template}, \link{az_vm_resource}, \link{az_vmss_template}, \link{az_vmss_resource}
}
