% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_methods.R
\name{list_vm_sizes}
\alias{list_vm_sizes}
\title{List available VM sizes}
\description{
Method for the \link[AzureRMR:az_subscription]{AzureRMR::az_subscription} class.
}
\section{Usage}{
\preformatted{list_vm_sizes(location, name_only = FALSE)
}
}

\section{Arguments}{

\itemize{
\item \code{location}: The location/region for which to obtain available VM sizes.
\item \code{name_only}: Whether to return only a vector of names, or all information on each VM size.
}
}

\section{Value}{

If \code{name_only} is TRUE, a character vector of names, suitable for passing to \code{create_vm}. If FALSE, a data frame containing the following information for each VM size: the name, number of cores, OS disk size, resource disk size, memory, and maximum data disks.
}

\seealso{
\link{create_vm}
}
