% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsg_config.R
\docType{data}
\name{nsg_rule_allow_ssh}
\alias{nsg_rule_allow_ssh}
\alias{nsg_rule_allow_http}
\alias{nsg_rule_allow_https}
\alias{nsg_rule_allow_rdp}
\alias{nsg_rule_allow_jupyter}
\alias{nsg_rule_allow_rstudio}
\alias{nsg_rule_allow_mssql}
\alias{nsg_rule_allow_mssql_browser}
\title{Network security rules}
\format{An object of class \code{nsg_rule} of length 2.}
\usage{
nsg_rule_allow_ssh

nsg_rule_allow_http

nsg_rule_allow_https

nsg_rule_allow_rdp

nsg_rule_allow_jupyter

nsg_rule_allow_rstudio

nsg_rule_allow_mssql

nsg_rule_allow_mssql_browser
}
\description{
Network security rules
}
\section{Usage}{
\preformatted{nsg_rule_allow_http
nsg_rule_allow_https
nsg_rule_allow_jupyter
nsg_rule_allow_rdp
nsg_rule_allow_rstudio
nsg_rule_allow_ssh
}
}

\section{Details}{

Some predefined network security rules provided by AzureVM, to unblock commonly-used ports.
\itemize{
\item HTTP: TCP port 80
\item HTTPS: TCP port 443
\item JupyterHub: TCP port 8000
\item RDP: TCP port 3389
\item RStudio Server: TCP port 8787
\item SSH: TCP port 22
\item SQL Server: TCP port 1433
\item SQL Server browser service: TCP port 1434
}
}

\keyword{datasets}
