% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_config.R
\name{vm_config}
\alias{vm_config}
\alias{ubuntu_dsvm}
\alias{windows_dsvm}
\alias{ubuntu_1604}
\alias{ubuntu_1804}
\alias{windows_2016}
\alias{windows_2019}
\alias{redhat_7.6}
\alias{redhat_8}
\alias{debian9_backports}
\title{VM configuration functions}
\usage{
vm_config(image, keylogin, managed = TRUE, datadisks = numeric(0),
  nsg = nsg_config(), ip = ip_config(), vnet = vnet_config(),
  nic = nic_config(), other_resources = list())

ubuntu_dsvm(keylogin = TRUE, managed = TRUE, datadisks = numeric(0),
  nsg = nsg_config(list(nsg_rule_allow_ssh, nsg_rule_allow_jupyter,
  nsg_rule_allow_rstudio)), ...)

windows_dsvm(keylogin = FALSE, managed = TRUE,
  datadisks = numeric(0), nsg = nsg_config(list(nsg_rule_allow_rdp)),
  ...)

ubuntu_1604(keylogin = TRUE, managed = TRUE, datadisks = numeric(0),
  nsg = nsg_config(list(nsg_rule_allow_ssh)), ...)

ubuntu_1804(keylogin = TRUE, managed = TRUE, datadisks = numeric(0),
  nsg = nsg_config(list(nsg_rule_allow_ssh)), ...)

windows_2016(keylogin = FALSE, managed = TRUE,
  datadisks = numeric(0), nsg = nsg_config(list(nsg_rule_allow_rdp)),
  ...)

windows_2019(keylogin = FALSE, managed = TRUE,
  datadisks = numeric(0), nsg = nsg_config(list(nsg_rule_allow_rdp)),
  ...)

redhat_7.6(keylogin = TRUE, managed = TRUE, datadisks = numeric(0),
  nsg = nsg_config(list(nsg_rule_allow_ssh)), ...)

redhat_8(keylogin = TRUE, managed = TRUE, datadisks = numeric(0),
  nsg = nsg_config(list(nsg_rule_allow_ssh)), ...)

debian9_backports(keylogin = TRUE, managed = TRUE,
  datadisks = numeric(0), nsg = nsg_config(list(nsg_rule_allow_ssh)),
  ...)
}
\arguments{
\item{image}{For \code{vm_config}, the VM image to deploy. Use \code{image_config} to specify this.}

\item{keylogin}{Boolean: whether to use an SSH public key to login (TRUE) or a password (FALSE). Note that Windows does not support SSH key logins.}

\item{managed}{Whether to provide a managed system identity for the VM.}

\item{datadisks}{The data disks to attach to the VM. Specify this as either a vector of numeric disk sizes in GB, or a list of \code{datadisk_config} objects for more control over the specification.}

\item{nsg}{The network security group for the VM.}

\item{ip}{The public IP address for the VM.}

\item{vnet}{The virtual network for the VM.}

\item{nic}{The network interface for the VM.}

\item{other_resources}{A list of other resources to include in the deployment.}
}
\description{
VM configuration functions
}
